initial buffer amplifier prototype test, v1.0
This is lifted from the kenwood ts-440s output buffer amplifier
on the IF board, which taps audio from the audio output before
the AF VR control, and provides a fixed level audio output to the
audio data output for things like digital modes.

earlier radios like the TS-670, TS-430, etc don't have a dedicated
13 pin accessory jack for digital modes and rig control.  this board
is for that.